requirements.txt
python-telegram-bot==13.10
gspread==3.6.0
oauth2client==4.1.3
pytest==6.0.1
pytest-mock==3.2.0